created an admin.ts in the controllers folder
